moved image grid 4 to its own component
