+++ src/components/ImageGrid2/_image-grid-2.scss
@@ -1,0 +1,23 @@
+.tco-image-grid-2 {
+  @include wider-than($breakpoint-tablet-portrait) {
+    display: flex;
+    flex-flow: row nowrap;
+    justify-content: space-between;
+  }
+
+  &-container {
+    width: 100%;
+
+    @include wider-than($breakpoint-tablet-portrait) {
+      width: calc(50% - #{$spacing-inline-16});
+    }
+
+    &:first-of-type {
+      margin-bottom: $spacing-stack-32;
+
+      @include wider-than($breakpoint-tablet-portrait) {
+        margin-bottom: 0;
+      }
+    }
+  }
+}
