+++ src/components/ImageGrid4/_image-grid-4.scss
@@ -1,27 +1,3 @@
-.tco-image-grid-2 {
-  @include wider-than($breakpoint-tablet-portrait) {
-    display: flex;
-    flex-flow: row nowrap;
-    justify-content: space-between;
-  }
-
-  &-container {
-    width: 100%;
-
-    @include wider-than($breakpoint-tablet-portrait) {
-      width: calc(50% - #{$spacing-inline-16});
-    }
-
-    &:first-of-type {
-      margin-bottom: $spacing-stack-32;
-
-      @include wider-than($breakpoint-tablet-portrait) {
-        margin-bottom: 0;
-      }
-    }
-  }
-}
-
 .tco-image-grid-4 {
   @include wider-than($breakpoint-phone-landscape) {
     display: flex;
