+++ src/components/_index.scss
@@ -13,7 +13,8 @@
 @import 'Forms/forms';
 @import 'ImageGrid/image-grid';
 @import 'ImageOnly/image-only';
-@import 'ImageGrid2-4/image-grid-2-4';
+@import 'ImageGrid2/image-grid-2';
+@import 'ImageGrid4/image-grid-4';
 @import 'Navigation/navigation';
 @import 'NewsletterForm/newsletter-form';
 @import 'PostInfo/post-info';
